move temp_local_dir to testing_utils
